requirements.txt
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-Login==0.5.0
itsdangerous==1.1.0
pymongo==3.11.0
python-dotenv==0.14.0
Werkzeug==1.0.1
